mysite/requirements.txt
asgiref==3.7.2
attrs==23.1.0
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.4
django-bootstrap5==23.3
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-filter==23.3
django-js-asset==2.1.0
django-summernote==0.8.20.0
django-taggit==4.0.0
django-taggit-serializer==0.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Markdown==3.4.4
oauthlib==3.2.2
Pillow==10.0.0
psycopg==3.1.10
psycopg-binary==3.1.10
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.3
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.5
webencodings==0.5.1
